Added PDF output. Added some python tests.
