+++ .gitignore
@@ -1,4 +1,4 @@
-.# compiled output
+# compiled output
 /dist/
 /declarations/
