+++ .vscode/settings.json
@@ -1,0 +1,15 @@
+{
+  "flake8.args": [
+   "--max-line-length=120",
+   "--ignore=E203,W503",
+],
+"python.testing.unittestArgs": [
+  "-v",
+  "-s",
+  "./pod/flask/python/",
+  "-p",
+  "*_test.py"
+],
+"python.testing.pytestEnabled": false,
+"python.testing.unittestEnabled": true,
+}
