+++ pod/flask/Containerfile
@@ -10,6 +10,14 @@ COPY requirements.txt requirements.txt
 # Install dependencies
 RUN pip install -r requirements.txt
 
+# Install Libreoffice
+RUN apt-get update -y
+RUN apt-get install auto-apt-proxy -y
+RUN apt-get install default-jre libreoffice-java-common libreoffice-writer-nogui libreoffice-draw-nogui -y
+
+# Copy the required font files to the image
+COPY ./fonts /usr/local/share/fonts
+
 # Copy the rest of the application code to the container
 COPY ./python .
