+++ pod/flask/python/Entry.py
@@ -3,22 +3,22 @@ from datetime import datetime
 
 class Entry(object):
     def __init__(
-      self,
-      id: 'int' = 0,
-      robotName: str = "",
-      coach: str = "",
-      school:  str = "",
-      competition: str = "",
-      driver1: str = "",
-      driver1Grade: int = 0,
-      driver2: str = "",
-      driver2Grade: int = 0,
-      driver3: str = "",
-      driver3Grade: int = 0,
-      checkInStatus: str = "",
-      measured: str = "",
-      registered: datetime = datetime.min,
-      paymentType: str = "UNPAID"
+        self,
+        id: "int" = 0,
+        robotName: str = "",
+        coach: str = "",
+        school: str = "",
+        competition: str = "",
+        driver1: str = "",
+        driver1Grade: int = 0,
+        driver2: str = "",
+        driver2Grade: int = 0,
+        driver3: str = "",
+        driver3Grade: int = 0,
+        checkInStatus: str = "",
+        measured: str = "",
+        registered: datetime = datetime.min,
+        paymentType: str = "UNPAID",
     ):
         self.id = id
         self.robotName = robotName
@@ -43,7 +43,7 @@ class Entry(object):
         return self.registered < other.registered
 
     def __cmp__(self, other):
-        if hasattr(other, 'robotName'):
+        if hasattr(other, "robotName"):
             return self.robotName.__cmp__(other.robotName)
         else:
             return False
