+++ pod/flask/python/EventEntry.py
@@ -1,7 +1,7 @@
 from Entry import Entry
 
 
-class EventEntry():
+class EventEntry:
     entry: Entry
     letter: str = ""
     position: int = -1
@@ -9,12 +9,12 @@ class EventEntry():
     placed: int = -1
 
     def __init__(
-      self,
-      entry: Entry,
-      letter: str = "",
-      position: int = -1,
-      placed: int = -1,
-    ) -> 'None':
+        self,
+        entry: Entry,
+        letter: str = "",
+        position: int = -1,
+        placed: int = -1,
+    ) -> "None":
         self.entry = entry
         self.letter = letter
         self.position = position
