+++ pod/flask/python/Resource.py
@@ -1,11 +1,12 @@
 """The Event object used by a Schedule"""
 
-
 # JUDGINGTIMEPERENTRY = datetime.timedelta(seconds=60)
 # FINALADJUDICATIONTIMEPERENTRY = datetime.timedelta(seconds=180)
 
+
 class Resource(object):
     """Used by a Schedule"""
+
     def __init__(self, name, lengthX, lengthY, resourceType):
         self.name = name
         self.lengthX = lengthX
