+++ pod/flask/python/RingAssignment.py
@@ -1,12 +1,12 @@
 class RingAssignment(object):
     def __init__(
-      self,
-      id: int = 0,
-      competition: str = "",
-      robot: int = -1,
-      letter: str = "",
-      placed: int = -1,
-      ring: int = -1,
+        self,
+        id: int = 0,
+        competition: str = "",
+        robot: int = -1,
+        letter: str = "",
+        placed: int = -1,
+        ring: int = -1,
     ):
         self.id = id
         self.competition = competition
