+++ pod/flask/python/api_test.py
@@ -1,0 +1,7 @@
+from api import api
+import unittest
+
+
+class testApiLoad(unittest.TestCase):
+    def test_api_load(self):
+        self.assertIsNotNone(api)
